refactor(LC011): split primary-key analysis helpers
- separate configuration scanning, type lookup, and key-rule analysis
- keep analyzer behavior unchanged while reducing local complexity
