+++ src/LinqContraband/Analyzers/SchemaAndModeling/LC011_EntityMissingPrimaryKey/EntityMissingPrimaryKeyConfigurationScan.cs
@@ -1,0 +1,131 @@
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace LinqContraband.Analyzers.LC011_EntityMissingPrimaryKey;
+
+public sealed partial class EntityMissingPrimaryKeyAnalyzer
+{
+    private static void ScanOnModelCreating(
+        INamedTypeSymbol dbContextType,
+        HashSet<INamedTypeSymbol> keylessEntities,
+        HashSet<INamedTypeSymbol> ownedEntities,
+        Compilation compilation)
+    {
+        var methods = dbContextType.GetMembers("OnModelCreating");
+        if (methods.IsEmpty || methods[0] is not IMethodSymbol onModelCreating)
+            return;
+
+        foreach (var syntaxRef in onModelCreating.DeclaringSyntaxReferences)
+        {
+            var syntax = syntaxRef.GetSyntax();
+
+            foreach (var invocation in syntax.DescendantNodes().OfType<InvocationExpressionSyntax>())
+            {
+                if (invocation.Expression is not MemberAccessExpressionSyntax memberAccess)
+                    continue;
+
+                var methodName = memberAccess.Name.Identifier.Text;
+                if (methodName == "HasNoKey")
+                    TryAddResolvedType(ExtractEntityTypeNameFromChain(memberAccess.Expression), compilation, keylessEntities);
+                else if (methodName is "OwnsOne" or "OwnsMany")
+                    TryAddResolvedType(ExtractOwnedTypeName(invocation), compilation, ownedEntities);
+            }
+        }
+    }
+
+    private static void ScanEntityTypeConfigurations(
+        Compilation compilation,
+        HashSet<INamedTypeSymbol> configuredEntities,
+        HashSet<INamedTypeSymbol> keylessEntities)
+    {
+        var configInterface = compilation.GetTypeByMetadataName("Microsoft.EntityFrameworkCore.IEntityTypeConfiguration`1");
+        if (configInterface == null)
+            return;
+
+        foreach (var type in GetAllTypes(compilation.Assembly.GlobalNamespace))
+        {
+            if (type.AllInterfaces.IsEmpty)
+                continue;
+
+            foreach (var iface in type.AllInterfaces)
+            {
+                if (!SymbolEqualityComparer.Default.Equals(iface.OriginalDefinition, configInterface))
+                    continue;
+
+                if (iface.TypeArguments.Length > 0 &&
+                    iface.TypeArguments[0] is INamedTypeSymbol entityType)
+                {
+                    var (hasKey, hasNoKey) = CheckConfigureMethod(type);
+
+                    if (hasKey)
+                        configuredEntities.Add(entityType);
+
+                    if (hasNoKey)
+                        keylessEntities.Add(entityType);
+                }
+            }
+        }
+    }
+
+    private static (bool hasKey, bool hasNoKey) CheckConfigureMethod(INamedTypeSymbol configClass)
+    {
+        var configureMethod = configClass.GetMembers("Configure").FirstOrDefault() as IMethodSymbol;
+        if (configureMethod == null)
+            return (false, false);
+
+        var hasKey = false;
+        var hasNoKey = false;
+
+        foreach (var syntaxRef in configureMethod.DeclaringSyntaxReferences)
+        {
+            var syntax = syntaxRef.GetSyntax();
+            var invocations = syntax.DescendantNodes().OfType<InvocationExpressionSyntax>();
+
+            foreach (var invocation in invocations)
+            {
+                if (invocation.Expression is not MemberAccessExpressionSyntax memberAccess)
+                    continue;
+
+                var methodName = memberAccess.Name.Identifier.Text;
+                if (methodName == "HasKey")
+                    hasKey = true;
+                if (methodName == "HasNoKey")
+                    hasNoKey = true;
+            }
+        }
+
+        return (hasKey, hasNoKey);
+    }
+
+    private bool HasFluentKeyConfiguration(INamedTypeSymbol dbContextType, INamedTypeSymbol entityType, Compilation compilation)
+    {
+        var methods = dbContextType.GetMembers("OnModelCreating");
+        if (methods.IsEmpty || methods[0] is not IMethodSymbol onModelCreating)
+            return false;
+
+        foreach (var syntaxRef in onModelCreating.DeclaringSyntaxReferences)
+        {
+            var syntax = syntaxRef.GetSyntax();
+
+            foreach (var invocation in syntax.DescendantNodes().OfType<InvocationExpressionSyntax>())
+            {
+                if (invocation.Expression is not MemberAccessExpressionSyntax memberAccess)
+                    continue;
+                if (memberAccess.Name.Identifier.Text != "HasKey")
+                    continue;
+
+                var entityTypeName = ExtractEntityTypeNameFromChain(memberAccess.Expression);
+                if (entityTypeName == null)
+                    continue;
+
+                var matchedType = FindTypeByName(compilation, entityTypeName);
+                if (matchedType != null && SymbolEqualityComparer.Default.Equals(matchedType, entityType))
+                    return true;
+            }
+        }
+
+        return false;
+    }
+}
